+++ python/cugraph/tests/test_balanced_cut.py
@@ -21,7 +21,6 @@ import cudf
 import cugraph
 from cugraph.tests import utils
 import rmm
-from rmm import rmm_config
 
 
 def cugraph_call(G, partitions):
@@ -57,11 +56,11 @@ PARTITIONS = [2, 4, 8]
 def test_modularity_clustering(managed, pool, graph_file, partitions):
     gc.collect()
 
-    rmm.finalize()
-    rmm_config.use_managed_memory = managed
-    rmm_config.use_pool_allocator = pool
-    rmm_config.initial_pool_size = 2 << 27
-    rmm.initialize()
+    rmm.reinitialize(
+        managed_memory=True,
+        pool_allocator=True,
+        initial_pool_size=2 << 27
+    )
 
     assert(rmm.is_initialized())