+++ python/cugraph/tests/test_filter_unreachable.py
@@ -21,7 +21,6 @@ import numpy as np
 import cugraph
 from cugraph.tests import utils
 import rmm
-from rmm import rmm_config
 
 # Temporarily suppress warnings till networkX fixes deprecation warnings
 # (Using or importing the ABCs from 'collections' instead of from
@@ -46,11 +45,11 @@ SOURCES = [1]
 def test_filter_unreachable(managed, pool, graph_file, source):
     gc.collect()
 
-    rmm.finalize()
-    rmm_config.use_managed_memory = managed
-    rmm_config.use_pool_allocator = pool
-    rmm_config.initial_pool_size = 2 << 27
-    rmm.initialize()
+    rmm.reinitialize(
+        managed_memory=True,
+        pool_allocator=True,
+        initial_pool_size=2 << 27
+    )
 
     assert(rmm.is_initialized())